+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to
 - Ability to create coordinates component with :ISO-X and :ISO-Y arguments.
 - Support for Tiled objects' type (point, ellipse, polygon, rectangle).
 - Ability to draw debug ellipses.
+- Movement target indicator for debugging.
 
 ### Changed
 - Changed syntax of WITH-SYSTEM-SLOTS macro to be more DSL-ish.